+++ src/app/api/(path)/topics/[id]/route.ts
@@ -2,6 +2,7 @@ import { Topic } from "@prisma/client";
 import { NextResponse } from "next/server";
 import prisma from "src/utils/prisma";
 import { revalidateTag } from "next/cache";
+import adminProtect from "src/app/api/utils/adminProtect";
 
 export async function GET(
   request: Request,
@@ -38,6 +39,8 @@ export async function PATCH(
   request: Request,
   { params }: { params: { id: string } }
 ) {
+  const resolved = await adminProtect();
+  if (!resolved) return NextResponse.json({ data: {} }, { status: 403 });
   const res = (await request.json()) as Partial<Topic>;
   const updateData = Object.fromEntries(
     Object.entries(res).map(([key, value]) => [key, { set: value }])
@@ -68,6 +71,8 @@ export async function DELETE(
   request: Request,
   { params: { id } }: { params: { id: string } }
 ) {
+  const resolved = await adminProtect();
+  if (!resolved) return NextResponse.json({ data: {} }, { status: 403 });
   const data = await prisma.topic.delete({
     where: { id: Number(id) },
     include: { Subject: { select: { Topics: { select: { id: true } } } } },
