+++ src/app/api/(users)/groups/[organization]/[id]/route.ts
@@ -3,12 +3,16 @@ import { Group, Role } from "@prisma/client";
 import { revalidateTag } from "next/cache";
 import { redirect } from "next/navigation";
 import { NextRequest, NextResponse } from "next/server";
+import organizationProtect from "src/app/api/utils/organizationProtect";
 import prisma from "src/utils/prisma";
 
 export async function PATCH(
   request: Request,
-  { params }: { params: { id: string } }
+  { params }: { params: { id: string, organization: string } }
 ) {
+  const organizationId = params?.organization
+  const resolve = organizationProtect({organization:organizationId});
+  if (!resolve) return NextResponse.json({ data: {} }, { status: 403 });
   const res = (await request.json()) as Partial<Group>;
   const updateData = Object.fromEntries(
     Object.entries(res).map(([key, value]) => [key, { set: value }])
@@ -49,6 +53,9 @@ export async function POST(
   request: Request,
   { params: { organization } }: { params: { organization: string } }
 ) {
+  const organizationId = organization
+  const resolve = organizationProtect({organization:organizationId});
+  if (!resolve) return NextResponse.json({ data: {} }, { status: 403 });
   const res = await request.json();
   const { name } = res;
   const data = await prisma.group.create({
@@ -59,8 +66,11 @@ export async function POST(
 
 export async function DELETE(
   request: Request,
-  { params }: { params: { id: string } }
+  { params }: { params: { id: string, organization: string } }
 ) {
+  const organizationId = params?.organization
+  const resolve = organizationProtect({organization:organizationId});
+  if (!resolve) return NextResponse.json({ data: {} }, { status: 403 });
   const id = Number(params.id);
   const data = await prisma.group.delete({
     where: {
