+++ src/app/api/(users)/organizations/[id]/route.ts
@@ -1,6 +1,7 @@
 import { Organization } from "@prisma/client";
 import { revalidateTag } from "next/cache";
 import { NextResponse } from "next/server";
+import adminProtect from "src/app/api/utils/adminProtect";
 import prisma from "src/utils/prisma";
 
 export async function GET(
@@ -27,6 +28,8 @@ export async function PATCH(
   request: Request,
   { params }: { params: { id: string } }
 ) {
+  const resolved = await adminProtect();
+  if (!resolved) return NextResponse.json({ data: {} }, { status: 403 });
   const res = (await request.json()) as Partial<Organization>;
   const updateData = Object.fromEntries(
     Object.entries(res).map(([key, value]) => [key, { set: value }])
@@ -52,6 +55,8 @@ export async function DELETE(
   request: Request,
   { params: { id } }: { params: { id: string } }
 ) {
+  const resolved = await adminProtect();
+  if (!resolved) return NextResponse.json({ data: {} }, { status: 403 });
   const data = await prisma.organization.delete({
     where: {
       id: Number(id),
